GH-2: Rename and reorganize Reddit resource classes.
