+++ app/src/main/java/com/syntactec/subreddify/resources/RedditResource.java
@@ -1,4 +1,4 @@
-package com.syntactec.subreddify.services;
+package com.syntactec.subreddify.resources;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
@@ -8,12 +8,9 @@ import retrofit2.http.Query;
 import java.util.List;
 
 /**
- * This class models the Reddit REST API.
+ * This class models the Reddit resources the application uses.
  */
-public interface RedditService {
+public interface RedditResource {
     @GET("r/{subreddit}/new.json?count=3")
     Call<List<RedditPost>> getPostsNewerThan(@Path("subreddit") String subreddit, @Query("before") String postName);
-
-    @GET("r/{subreddit}/comments.json?count=3")
-    Call<List<RedditComment>> getCommentsNewerThan(@Path("subreddit") String subreddit, @Query("before") String postName);
 }
